+++ public_html/wp-content/themes/mbd-wp-theme/functions.php
@@ -1,11 +1,22 @@
 <?php
 
 /**
- * Load the theme framework
+ * Load the custom theme if it exists.
+ * 
+ * By loading this first, this allow us to 
+ * override framework functions.
+ * 
+ */
+include get_template_directory() . '/custom/custom.php';
+
+/**
+ * Load the theme framework.
+ * 
  */
 require get_template_directory() . '/framework/framework.php';
 
 /**
- * Load the custom theme if it exists
+ * Remove framework hooks as required.
+ * 
  */
-include get_template_directory() . '/custom/custom.php';
+include get_template_directory() . '/custom/remove-hooks.php';